% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_all}
\alias{data_all}
\title{Data used to train the model.}
\format{A data frame with 2000 rows and 10 variables:
\describe{
  \item{X1}{binary, 0 or 1}
  \item{X2}{binary, 0 or 1}
  \item{X3}{binary, 0 or 1}
  \item{X4}{binary, 0 or 1}
  \item{X5}{binary, 0 or 1}
  \item{Y}{dependent variable, continuous}
  \item{IV}{instrumental variable, continuous}
  ...
}}
\source{
The data is generated by data_generator.R in data-raw directory
}
\usage{
data_all
}
\description{
A dataset containing five X variables and other auxiliary variables
}
\keyword{datasets}
