% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest.IVTree.R
\name{honest.IVTree}
\alias{honest.IVTree}
\title{...}
\usage{
honest.IVTree(
  formula,
  data,
  weights,
  treatment,
  treatment1,
  IV,
  subset,
  est_data,
  est_weights,
  est_treatment,
  est_treatment1,
  est_IV,
  est_subset,
  na.action = na.IVTree,
  split.Honest,
  HonestSampleSize,
  split.Bucket,
  bucketNum = 10,
  bucketMax = 40,
  cv.Honest,
  cv.option,
  minsize = 2L,
  model = FALSE,
  x = FALSE,
  y = TRUE,
  propensity,
  control,
  split.alpha = 0.5,
  cv.alpha = 0.5,
  cv.gamma = 0.5,
  split.gamma = 0.5,
  cost,
  ...
)
}
\arguments{
\item{...}{....}
}
\value{
....
}
\description{
...
}
